+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+#Version 0.5.0
+
+##Features
+
+- Add handler on change to checkbox 
+- Add handler on change to elements marked form-control class
+- Add handler on change to file input elements marked form-control-wrapper class
+- Add handler on key down to elements marked form-control class
+- Add handler on past to elements marked form-control class
+- Add handler on key up to elements marked form-control class
+- Add handler on focus to file input elements marked form-control-wrapper class
+- Add handler on blur to file input elements marked form-control-wrapper class
+
 #Version 0.4.0
 
 ##Features